llmspell_core

Module events

ABOUTME: Event types and structures for llmspell-core ABOUTME: Provides artifact lifecycle events and event metadata

Re-exports

pub use artifact_events::AccessType;

pub use artifact_events::ArtifactAccessedEvent;

pub use artifact_events::ArtifactCreatedEvent;

pub use artifact_events::ArtifactDeletedEvent;

pub use artifact_events::ArtifactDerivedEvent;

pub use artifact_events::ArtifactEvent;

pub use artifact_events::ArtifactEventBuilder;

pub use artifact_events::ArtifactEventType;

pub use artifact_events::ArtifactMetadataUpdatedEvent;

pub use artifact_events::ArtifactModifiedEvent;

pub use artifact_events::ArtifactValidatedEvent;

pub use artifact_events::ArtifactValidationFailedEvent;

pub use artifact_events::ArtifactVersionedEvent;

pub use artifact_events::DerivationType;

pub use artifact_events::Modification;

pub use artifact_events::StorageLocation;

pub use artifact_events::ValidationFailure;

pub use artifact_events::ValidationResults;

pub use artifact_events::ValidationType;

Modules

artifact_events

ABOUTME: Event types and structures for artifact lifecycle and operations ABOUTME: Provides events for artifact creation, modification, deletion, and correlation