llmspell_core::traits

Module tool_capable

ABOUTME: ToolCapable trait for components that can interact with tools ABOUTME: Separate trait to maintain clean architecture and avoid trait cyclicity

Structs

ToolComposition

Tool composition definition.

ToolCompositionStep

Tool composition step for executing multiple tools in sequence.

ToolInfo

Information about a tool that can be invoked.

ToolQuery

Tool query criteria for tool discovery.

Enums

ContextMode

Context passing mode for tool composition.

ErrorStrategy

Error handling strategy for tool composition steps.

Traits

ToolCapable

Tool-capable trait for components that can interact with tools.